backend/requirements.txt
django>=4.2,<5.0
django-environ
djangorestframework
django-cors-headers
pyjwt
python-jose[cryptography]
requests
psycopg2-binary
boto3
django-storages